Adding controller to site structure
