+++ controllers/error.js
@@ -1,0 +1,3 @@
+exports.get404 = (req, res, next) => {
+    res.status(404).render('404', { pageTitle: 'Page Not Found' });
+};
