+++ views/404.ejs
@@ -2,7 +2,8 @@
 </head>
 
 <body>
-    <%- include('includes/navigation.ejs') %> 
-    <h1>Page Not Found!</h1>
+    <main>
+        <h1>Page Not Found!</h1>
+    </main>
 
 <%- include('includes/end.ejs') %>
